[gas] trivial: move StorageGasParameters to its own module
